master: implement a loop to create meta/trigger builders
This reduces code duplication and makes it easier to add a force
scheduler:

When running a local buildbot, it makes sense to support force
triggering of build. Support this by adding force schedulers for each
arch/machine when the MASTER_ENABLE_FORCE environment variable is
defined.

To enable the force schedulers, add '-e MASTER_ENABLE_FORCE=1' to the
list of arguments to the 'docker run' command, and login through the web
UI to allow force builds.