master: disable logging of environment
The slaves are configured through the environment when running the
docker container. Since the environment is logged to console by default,
it exposes the slave passwords to the world.

Avoid this by disabling the logging of the env. Note that later changes
to master.cfg need to be carefull to disable logEnviron when adding
steps.