master: fix max_builds for slaves
the buildbot code is expecting this to be an integer value with no
warnings/asserts to check it at configuration time. Before this
change the code would do a "if len(current_builds) >= max_builds"
check that will always return false when max_builds is a string.

you fail it Python!